【修改】修改大部分界面

+++ src/com/scnu/swimmingtrainingsystem/activity/TimerActivity.java
@@ -105,7 +105,7 @@ public class TimerActivity extends Activity {
 		app = (MyApplication) getApplication();
 		// 如果app中的全局变量被系统强制回收，通过以下改行代码会触发异常，直接将应用界面重启至登陆页面
 		@SuppressWarnings("unused")
-		long userID = (Long) app.getMap().get(Constants.CURRENT_USER_ID);
+		int userID = (Integer) app.getMap().get(Constants.CURRENT_USER_ID);
 		time_title = (TextView) findViewById(R.id.time_title);
 		tvTime = (TextView) findViewById(R.id.duocitvTime);
 		tvTip = (TextView) findViewById(R.id.textwujici);
